[Cleanup] Config: moved away helper functions
config.h/cpp: Helper functions moved to util.h/cpp. Server iterfacing function moved to messaging.h/cpp.
config.h: #include <rudeconfig/config.h> moved to .cpp
Removed header guard macros, kept #pragma once
