+++ source/server/messaging.h
@@ -17,8 +17,6 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 #pragma once
-#ifndef __Messaging_H__
-#define __Messaging_H__
 
 class SWInetSocket;
 
@@ -44,10 +42,11 @@ public:
 	static void updateMinuteStats();
 	static int getTime();
 
+	static std::string retrievePublicIpFromServer();
+
 protected:
 	static stream_traffic_t traffic;
 };
 
 
-#endif
